frontend: refactor to separate a select and non-select explorers
