+++ frontend/src/components/TagExplorer.tsx
@@ -2,90 +2,37 @@
 import { Button, Stack, Typography } from "@mui/joy";
 import { RichTreeView, TreeViewBaseItem } from "@mui/x-tree-view";
 import React, { FC, useEffect, useState } from "react";
+import { useNavigate } from "react-router";
 import {
-  ReadTagsByFileIDsResponse,
   Tag,
   TagService,
 } from "../../bindings/github.com/michael-freling/anime-image-viewer/internal/image";
-import {
-  ExplorerTreeItem,
-  ExplorerTreeItemLabel,
-  ExplorerTreeItemProps,
-} from "./ExplorerTreeItem";
-import { useNavigate } from "react-router";
-
-export type TagExplorerProps =
-  | {
-      title: string;
-      editable: boolean;
-    }
-  | {
-      selectable: boolean;
-      onSelect: (addedTagIds: number[], deletedTagIds: number[]) => void;
-      fileIds: number[];
-    };
+import { ExplorerTreeItem, ExplorerTreeItemProps } from "./ExplorerTreeItem";
 
 const tagsToTreeViewBaseItems = (
   tags: Tag[],
-  fileCount: number,
-  addedTagIds: { [key: number]: boolean },
-  deletedTagIds: { [key: number]: boolean },
-  tagStats?: ReadTagsByFileIDsResponse
+  fileCount: number
 ): TreeViewBaseItem<{
   id: string;
   label: string;
-  count: number;
-  indeterminate: boolean;
-  checked: boolean;
 }>[] => {
   return tags.map((child) => {
-    let count = 0;
-    if (tagStats != undefined && tagStats.TagCounts[child.ID] > 0) {
-      count = tagStats.TagCounts[child.ID];
-    }
-    let disabled = false;
-    if (tagStats != undefined && tagStats.AncestorMap[child.ID] != undefined) {
-      disabled = true;
-    }
-
-    const isAdded = addedTagIds[child.ID];
-    const isDeleted = deletedTagIds[child.ID];
-
     return {
-      id: String(child.ID),
-      label: child.Name,
-      children: tagsToTreeViewBaseItems(
-        child.Children,
-        fileCount,
-        addedTagIds,
-        deletedTagIds,
-        tagStats
-      ),
-      count,
-      indeterminate:
-        isAdded == undefined &&
-        isDeleted == undefined &&
-        count > 0 &&
-        count < fileCount,
-      checked: isAdded || count == fileCount,
-      disabled,
+      id: String(child.id),
+      label: child.name,
+      children: tagsToTreeViewBaseItems(child.children ?? [], fileCount),
     };
   });
 };
 
+export interface TagExplorerProps {
+  title: string;
+  editable: boolean;
+}
 const TagExplorer: FC<TagExplorerProps> = (props) => {
   const navigate = useNavigate();
   const [children, setChildren] = useState<Tag[]>([]);
-  const [tagStats, setTagStats] = useState<ReadTagsByFileIDsResponse>();
-  const [tagStatsLoaded, setTagStatsLoaded] = useState(false);
-  const [addedTagIds, setAddedTagIds] = useState<{ [key: number]: boolean }>(
-    {}
-  );
-  const [deletedTagIds, setDeletedTagIds] = useState<{
-    [key: number]: boolean;
-  }>({});
 
-  const selectable = "selectable" in props;
   useEffect(() => {
     if (children.length > 0) {
       return;
@@ -97,122 +44,13 @@ const TagExplorer: FC<TagExplorerProps> = (props) => {
   async function refresh() {
     const tags = await TagService.GetAll();
     setChildren(tags);
-    // setMap(getTagMap(tags));
-    if (tagStatsLoaded) {
-      return;
-    }
-
-    if (selectable) {
-      const fileIds = props.fileIds;
-      const response = await TagService.ReadTagsByFileIDs(fileIds);
-      setTagStats(response);
-      setTagStatsLoaded(true);
-    }
   }
 
   if (children.length === 0) {
     return <Typography>Loading...</Typography>;
   }
 
-  const treeItems = tagsToTreeViewBaseItems(
-    children,
-    "fileIds" in props ? props.fileIds.length : 0,
-    addedTagIds,
-    deletedTagIds,
-    tagStats
-  );
-
-  if (selectable) {
-    if (!tagStatsLoaded) {
-      return <Typography>Loading...</Typography>;
-    }
-
-    const { onSelect } = props;
-    const getAllTreeItemIds = (
-      items: TreeViewBaseItem<{
-        id: string;
-      }>[]
-    ): string[] => {
-      const itemIds: string[] = [];
-      for (let item of items) {
-        itemIds.push(item.id);
-        if (!item.children) {
-          continue;
-        }
-        itemIds.push(...getAllTreeItemIds(item.children));
-      }
-
-      return itemIds;
-    };
-
-    const defaultExpandedItems = getAllTreeItemIds(treeItems);
-    const defaultSelectedItems = Object.keys(tagStats!.TagCounts).filter(
-      (tagId) => {
-        return tagStats?.TagCounts[tagId] == props.fileIds.length;
-      }
-    );
-    return (
-      <RichTreeView
-        sx={{
-          flexGrow: 1,
-        }}
-        expansionTrigger="content"
-        defaultExpandedItems={defaultExpandedItems}
-        defaultSelectedItems={defaultSelectedItems}
-        slots={{
-          // todo: RichTreeView doesn't allow to pass a type other than TreeItem2Props
-          item: ExplorerTreeItem as any,
-        }}
-        slotProps={{
-          item: {
-            labelComponent: ExplorerTreeItemLabel,
-            selectable: true,
-          } as ExplorerTreeItemProps,
-        }}
-        onSelectedItemsChange={(
-          event: React.SyntheticEvent,
-          tagIds: string[]
-        ) => {
-          if (!tagIds) {
-            return;
-          }
-          let initialSelectedTagIds: string[] = [];
-          let initialAllTagIds: string[] = [];
-          if (tagStats) {
-            for (let [tagId, count] of Object.entries(tagStats.TagCounts)) {
-              if (count == props.fileIds.length) {
-                initialSelectedTagIds.push(tagId);
-              }
-              initialAllTagIds.push(tagId);
-            }
-          }
-
-          const addedTagIds = tagIds
-            .filter((tagId) => !initialSelectedTagIds.includes(tagId))
-            .map((tagId) => parseInt(tagId, 10));
-          const deletedTagIds = initialAllTagIds
-            .filter((tagId) => !tagIds.includes(tagId))
-            .map((tagId) => parseInt(tagId, 10));
-          setAddedTagIds(
-            addedTagIds.reduce((acc, tagId) => {
-              acc[tagId] = true;
-              return acc;
-            }, {} as { [key: number]: boolean })
-          );
-          setDeletedTagIds(
-            deletedTagIds.reduce((acc, tagId) => {
-              acc[tagId] = true;
-              return acc;
-            }, {} as { [key: number]: boolean })
-          );
-          onSelect(addedTagIds, deletedTagIds);
-        }}
-        items={treeItems}
-        multiSelect={true}
-        checkboxSelection={true}
-      />
-    );
-  }
+  const treeItems = tagsToTreeViewBaseItems(children, 0);
 
   const { title, editable } = props;
   if (editable) {
@@ -268,8 +106,6 @@ const TagExplorer: FC<TagExplorerProps> = (props) => {
           slotProps={{
             item: {
               addNewChild,
-              labelComponent: ExplorerTreeItemLabel,
-              selectable: false,
             } as ExplorerTreeItemProps,
           }}
           isItemEditable={() => true}
@@ -305,12 +141,6 @@ const TagExplorer: FC<TagExplorerProps> = (props) => {
         slots={{
           // todo: RichTreeView doesn't allow to pass a type other than TreeItem2Props
           item: ExplorerTreeItem as any,
-        }}
-        slotProps={{
-          item: {
-            labelComponent: ExplorerTreeItemLabel,
-            selectable: false,
-          } as ExplorerTreeItemProps,
         }}
         onSelectedItemsChange={(
           event: React.SyntheticEvent,
