+++ frontend/src/pages/DirectoryTagEditPage.tsx
@@ -2,7 +2,7 @@ import { Box, Button, Stack, Typography } from "@mui/joy";
 import { useState } from "react";
 import { useNavigate, useSearchParams } from "react-router";
 import { TagService } from "../../bindings/github.com/michael-freling/anime-image-viewer/internal/image";
-import TagExplorer from "../components/TagExplorer";
+import SelectTagExplorer from "../components/SelectTagExplorer";
 
 const DirectoryTagsEditPage = () => {
   const navigate = useNavigate();
@@ -32,8 +32,7 @@ const DirectoryTagsEditPage = () => {
       }}
     >
       <Typography>Select tags for {directoryIds.length} directories</Typography>
-      <TagExplorer
-        selectable={true}
+      <SelectTagExplorer
         fileIds={directoryIds}
         onSelect={(addedTagIds, deletedTagIds) => {
           setAddedTagIds(addedTagIds);
